DataContext güncellendi

+++ Backend/GuneyTeletip/GT.Persistance.Domain/Models/InfStudyParameter.cs
@@ -9,7 +9,7 @@ namespace GT.Persistance.Domain.Models
         public DateTime? TimeStart { get; set; }
         public DateTime? TimeStop { get; set; }
         public long? OracleStudyKeyLast { get; set; }
-        public short? RecordStatus { get; set; }
+        public long? RecordStatus { get; set; }
         public DateTime? TimeCreated { get; set; }
         public DateTime? TimeModified { get; set; }
         public long? FkUserCreated { get; set; }
